fix(hunyuan-video10): keep scaled timestep in fp32

+++ unirl/models/hunyuan_video10/diffusion.py
@@ -60,8 +60,8 @@ class HunyuanVideo10DiffusionStep(DiffusionStep[HunyuanVideo10Bundle, HunyuanVid
             timestep = sigma.expand(batch_size)
         else:
             timestep = sigma
-        # Match Diffusers/SGLang: scale fp32 sigma before casting to the transformer dtype.
-        timestep = (timestep.to(device=device, dtype=torch.float32) * self.TIMESTEP_SCALE).to(dtype=dtype)
+        # Match Diffusers/SGLang: keep the scaled timestep in fp32 through the sinusoidal embedding.
+        timestep = timestep.to(device=device, dtype=torch.float32) * self.TIMESTEP_SCALE
 
         guidance = torch.full((batch_size,), guidance_scale, device=device, dtype=dtype)
